requirements.txt
groq
python-dotenv==1.0.1
pandas==2.2.2
streamlit